Migrate other queries endpoints
